fix: Always capture stdout and remove syncing Sentry test repo (#47)
Multiple executions of the app can occur on the same container. Unfortunately, we were syncing Sentry to the *test* repo upon start up and having a Git lock. Instead of investigating it, we can simply rely on the next `git push` event on Sentry to sync both repos (since an event is only handled by one thread).

Subprocess calls need their output captured in order to show in Sentry's error reporting. Capturing the output will have the side-effect of making a human have to wait for the output of a command (which is acceptable).