feat: Only return failure status code when there's a failure in the execution
Most events from Github's webhooks events are not to be processed (that's why
we were returning something different than 200), yet, because of it's bad UI
we will return 200 by default and something different if something fails.

On the other hand, Git revert requests are triggered by a human and will follow
the standard status code responses.