feat: Booting code to report errors to Sentry
When an error is not caught at the top level, the worker is only given a bit
of time to report the error before shutting down.

In order to report the error to Sentry we need to capture first, report it
and then re-raise it for the worker to fail and shut down.