feat(devenv): Get rid of docker-compose and other changes
This initially started with me investigating `docker-compose build` including the contents of `__pycache__`.
I thought I managed to make `docker build` work but not the image build by `docker-compose`.
I believed that docker-compose was not respecting the build context (because we define a different one in
the docker-compose file) and because we mount the source checkout.
Since we have a working set-up with `flask run`, let's get rid of docker-compose since it can sometimes be
burdensome to troubleshoot.

Other changes:
* Make the Dockerfile only copy `*py` files, thus, avoding to include `__pycache__`
* Initial direnv execution to install Python dependencies
* Fix deploy workflow to consider changes under the `gitbot` rather than specific files
* Update docs to remove mention of `docker-compose`
* Some code changes to permit local development without a Personal Access Token
* Include `black` as dev dependency since VS Code keeps asking for it and is part of our pre-commit set up